# ride/internal/adapters/http/api/types.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.8.0 DO NOT EDIT.
package api

import (
	"time"
)

// Defines values for AssignmentStatus.
const (
	AssignmentStatusActive    AssignmentStatus = "active"
	AssignmentStatusCompleted AssignmentStatus = "completed"
	AssignmentStatusPending   AssignmentStatus = "pending"
)

// Valid indicates whether the value is a known member of the AssignmentStatus enum.
func (e AssignmentStatus) Valid() bool {
	switch e {
	case AssignmentStatusActive:
		return true
	case AssignmentStatusCompleted:
		return true
	case AssignmentStatusPending:
		return true
	default:
		return false
	}
}

// Defines values for ListAssignmentsParamsStatus.
const (
	ListAssignmentsParamsStatusActive    ListAssignmentsParamsStatus = "active"
	ListAssignmentsParamsStatusCompleted ListAssignmentsParamsStatus = "completed"
	ListAssignmentsParamsStatusPending   ListAssignmentsParamsStatus = "pending"
)

// Valid indicates whether the value is a known member of the ListAssignmentsParamsStatus enum.
func (e ListAssignmentsParamsStatus) Valid() bool {
	switch e {
	case ListAssignmentsParamsStatusActive:
		return true
	case ListAssignmentsParamsStatusCompleted:
		return true
	case ListAssignmentsParamsStatusPending:
		return true
	default:
		return false
	}
}

// Assignment defines model for Assignment.
type Assignment struct {
	Metadata  *EntityMetadata  `json:"metadata,omitempty"`
	RouteId   string           `json:"routeId"`
	StartsAt  time.Time        `json:"startsAt"`
	Status    AssignmentStatus `json:"status"`
	VehicleId string           `json:"vehicleId"`
}

// AssignmentStatus defines model for Assignment.Status.
type AssignmentStatus string

// EntityMetadata defines model for EntityMetadata.
type EntityMetadata struct {
	CreatedAt *time.Time `json:"createdAt,omitempty"`
	Id        *string    `json:"id,omitempty"`
	UpdatedAt *time.Time `json:"updatedAt,omitempty"`
}

// NewAssignment defines model for NewAssignment.
type NewAssignment struct {
	RouteId   string    `json:"routeId"`
	StartsAt  time.Time `json:"startsAt"`
	VehicleId string    `json:"vehicleId"`
}

// BadRequest defines model for BadRequest.
type BadRequest struct {
	Details *string `json:"details,omitempty"`
	Error   *string `json:"error,omitempty"`
}

// NotFound defines model for NotFound.
type NotFound struct {
	Details *string `json:"details,omitempty"`
	Error   *string `json:"error,omitempty"`
}

// ListAssignmentsParams defines parameters for ListAssignments.
type ListAssignmentsParams struct {
	Status *ListAssignmentsParamsStatus `form:"status,omitempty" json:"status,omitempty"`
}

// ListAssignmentsParamsStatus defines parameters for ListAssignments.
type ListAssignmentsParamsStatus string

// CreateAssignmentJSONRequestBody defines body for CreateAssignment for application/json ContentType.
type CreateAssignmentJSONRequestBody = NewAssignment